Add mask support for masked binary operations - c2 only changes
This patch adds the mask support for binary operations by transforming
the original "blend" pattern to the supported masked IR for platforms
that supports mask feature like SVE.

For example, the following change will be applied to masked vector add:

  VectorBlend (src1 (AddVI src1 src2) mask)  ==> AddVI (src1 src2 mask)

It doesn't change the Vector API java implementations. Instead, most of
the changes focus on the C2 compiler mid-end part. The mask value will
be used as an additional input to the original binary vector.