Add mask support for masked vector store
Currently masked vector store is not vectorized, since it only
needs the active elements being stored to the memory. This is
not vectorizable for platforms that do not have mask features.
However, it can be vectorized with a mask controlling for the
mask supported platforms like SVE.

This patch makes it possible to vectorize the masked vector
store by adding an intrinsic named "storeMasked" and the hotspot
implementation.

The java implementation of the masked store API will call the
new added intrinisc directly. And the hotspot needs to check
whether the masked store is supported and implemented for the
current platform. Finally it generates the masked store IR if all
the conditions are matched, otherwise it returns to the default
java implementation.

TODO: Add the support for storing values to the byte array/buffer
while the current element type is not byte. Besides, except for
the missing support for arithmetic operations, the masked load
and gather_load/scatter_store are also in the TODO list.